+++ src/features/categories/delete/DeleteConfirmationDialog.tsx
