+++ src/features/inventory/FilterBar.tsx
