+++ src/features/inventory/InventoryTabs.tsx
