+++ src/features/inventory/delete/DeleteConfirmationDialog.tsx
