+++ src/features/inventory/delete/DeleteProductButton.tsx
