+++ src/features/supplies/delete/DeleteButton.tsx
