+++ src/features/supplies/delete/DeleteConfirmationDialog.tsx
