+++ src/features/variants/delete/DeleteButton.tsx
