+++ src/features/variants/delete/DeleteConfirmationDialog.tsx
